feat(checkbox): 代码优化 样式调整
