+++ packages/yike-design-ui/src/components/checkbox/src/checkbox.ts
@@ -4,16 +4,3 @@ export type CheckboxProps = {
   value?: string | number | boolean;
   disabled?: boolean;
 };
-
-export const toBoolean = (v: number | string | boolean) => {
-  if (typeof v === 'boolean') {
-    return v;
-  }
-  return !!v.toString();
-};
-export const isArray = (v: any): v is any[] => {
-  return Array.isArray(v);
-};
-export const isUndefined = (v: any) => {
-  return typeof v === 'undefined';
-};
